+++ tests/core/simulation_control_02.cc
@@ -24,13 +24,14 @@ test()
   simulationControlParameters.method =
     Parameters::SimulationControl::TimeSteppingMethod::bdf1;
 
-  simulationControlParameters.timeEnd                = 999;
-  simulationControlParameters.number_mesh_adaptation = 9;
-  simulationControlParameters.output_name            = "test";
-  simulationControlParameters.subdivision            = 7;
-  simulationControlParameters.output_folder          = "canard";
-  simulationControlParameters.output_frequency       = 8;
-  simulationControlParameters.output_time_interval   = {0, 1000000000};
+  simulationControlParameters.time_end                   = 999;
+  simulationControlParameters.number_mesh_adaptation     = 9;
+  simulationControlParameters.output_name                = "test";
+  simulationControlParameters.subdivision                = 7;
+  simulationControlParameters.output_folder              = "canard";
+  simulationControlParameters.output_iteration_frequency = 8;
+  simulationControlParameters.output_time_interval       = {0, 1000000000};
+  simulationControlParameters.time_step_independent_of_end_time = true;
 
   SimulationControlSteady simulation_control(simulationControlParameters);