+++ tests/core/specific_heat_phase_change_01.cc
@@ -44,13 +44,13 @@ test()
   simulationControlParameters.maxCFL = 99;
   simulationControlParameters.method =
     Parameters::SimulationControl::TimeSteppingMethod::bdf1;
-  simulationControlParameters.timeEnd                = 999;
-  simulationControlParameters.number_mesh_adaptation = 9;
-  simulationControlParameters.output_name            = "test";
-  simulationControlParameters.subdivision            = 7;
-  simulationControlParameters.output_folder          = "canard";
-  simulationControlParameters.output_frequency       = 8;
-
+  simulationControlParameters.time_end                   = 999;
+  simulationControlParameters.number_mesh_adaptation     = 9;
+  simulationControlParameters.output_name                = "test";
+  simulationControlParameters.subdivision                = 7;
+  simulationControlParameters.output_folder              = "canard";
+  simulationControlParameters.output_iteration_frequency = 8;
+  simulationControlParameters.output_time_interval = {0, 1.7976931348623157e3};
 
   std::shared_ptr<SimulationControl> simulation_control;
   simulation_control =